wip: consolidate types
